django: patch_db_connections: fix RuntimeError
When patch_db_connections() is used together with aiohttp-wsgi,
asyncio.Task.get_current_task() doesn't return None as expected but
raises a RuntimeError.

The RuntimeError says there is no event loop running in 'Thread-${NUMBER}'
which is correct, so we return the original thread safe Django connections to
have the standard transaction behavior in the wsgi container.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>