rpc: add TreadedWorkerPool
In the future this code will run sync code like json.*, all protocol code and
Django code for example in a ThreadPoolExecutor to avoid blocking the io loop.

This code is currently disabled by setting the max_workers to 0 but i submit it
now to test it and use it explicit in another project.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>